fixing init and sandbox so the app runs again, incorporating PR #40 because I couldn't stop myself (and to make sure it works)
