+++ app/helpers.py
@@ -47,6 +47,57 @@ def prepare_transactions_for_split(trip_id: int) -> List[List[Union[str, float]]
     db.close()
     return result
 
+def normalize_name(name, name_lookup=None):
+    """
+    Normalize a name consistently.
+    If name exists in lookup, use that canonical version.
+    Otherwise, properly capitalize the name.
+    """
+    # Clean the input
+    cleaned_name = name.strip()
+    if not cleaned_name:
+        return cleaned_name
+        
+    # Check for existing canonical version
+    if name_lookup:
+        lower_name = cleaned_name.lower()
+        if lower_name in name_lookup:
+            return name_lookup[lower_name]
+    
+    # For new names, ensure consistent capitalization
+    # (capitalize first letter of each word)
+    return " ".join(word.capitalize() for word in cleaned_name.split())
+
+def update_session_names(db, session, trip_id):
+    """
+    Completely rebuild the session name lists from the database
+    to ensure consistency.
+    """
+    # Query all unique participants and payers from the database for this trip
+    participants_query = (
+        db.query(Participant.name)
+        .join(Transaction, Participant.trans_id == Transaction.trans_id)
+        .filter(Transaction.trip_id == trip_id)
+        .distinct()
+    )
+    
+    payers_query = (
+        db.query(Participant.name)
+        .join(Transaction, Participant.trans_id == Transaction.trans_id)
+        .filter(Transaction.trip_id == trip_id, Participant.is_payer == True)
+        .distinct()
+    )
+    
+    # Get results as lists of names
+    all_participants = [p[0] for p in participants_query.all()]
+    all_payers = [p[0] for p in payers_query.all()]
+    
+    # Update session with fresh data
+    session["unique_participants"] = all_participants
+    session["unique_payers"] = all_payers
+    
+    # Return the name lists for immediate use
+    return all_participants, all_payers
 
 # generate artificial transactions for testing the algorithm 
 def generate_transactions(num_people, num_transactions_per_person=10, max_amount=100):
